requirements.txt
playwright==1.42.0
playwright-stealth==1.0.6
anthropic==0.20.0
openai==1.35.7
pyyaml==6.0.1
schedule==1.2.1
flask==3.0.2
flask-cors==4.0.0
flask-socketio==5.3.6
aiohttp==3.9.3
python-dotenv==1.0.1
requests==2.31.0
